requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
openpyxl==3.1.2
reportlab==4.0.7
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
python-dateutil==2.8.2
pytz==2023.3
gunicorn==21.2.0
whitenoise==6.6.0

